fix(converter): accept redact param in standalone ToJSON/ToYAML and address review findings
- Add `redact bool` parameter to JSONConverter.ToJSON and
  YAMLConverter.ToYAML so callers can opt into sensitive field
  redaction without going through HybridGenerator.
- Move loadTestDataFromFile from markdown_integration_test.go
  to test_helpers.go for shared test access.
- Rename TestNewFieldsSerialization to RunNewFieldsSerializationTests
  (was in non-_test.go file, never executed) and wire it into
  enrichment_test.go with redact=true so assertions actually run.
- Add CRLF→LF normalization in normalizeSerializedOutput for
  cross-platform golden file comparison.
- Add POSIX trailing newlines to all golden JSON and MD files.
- Update docs/api.md examples with new signature.
