+++ internal/converter/testdata/golden/complete_redacted.golden.json
