+++ internal/converter/testdata/golden/complete_unredacted.golden.json
