+++ internal/converter/testdata/golden/edge_cases_redacted.golden.json
