+++ internal/converter/testdata/golden/edge_cases_unredacted.golden.json
