+++ internal/converter/testdata/golden/minimal_redacted.golden.json
