+++ internal/converter/testdata/golden/minimal_unredacted.golden.json
